+++ templates/shared_services/admin-vm/terraform/admin-jumpbox.tf
@@ -37,11 +37,16 @@ resource "azurerm_windows_virtual_machine" "jumpbox" {
   admin_password             = random_password.password.result
   tags                       = local.tre_shared_service_tags
 
-  source_image_reference {
-    publisher = "MicrosoftWindowsDesktop"
-    offer     = "windows-11"
-    sku       = "win11-24h2-pro"
-    version   = "latest"
+  # set source_image_id/reference depending on the config for the selected image
+  source_image_id = local.selected_image_source_id
+  dynamic "source_image_reference" {
+    for_each = local.selected_image_source_refs
+    content {
+      publisher = source_image_reference.value["publisher"]
+      offer     = source_image_reference.value["offer"]
+      sku       = source_image_reference.value["sku"]
+      version   = source_image_reference.value["version"]
+    }
   }
 
   os_disk {