+++ templates/shared_services/admin-vm/terraform/variables.tf
@@ -20,3 +20,12 @@ variable "enable_cmk_encryption" {
 variable "key_store_id" {
   type = string
 }
+
+variable "image_gallery_id" {
+  type    = string
+  default = ""
+}
+
+variable "image" {
+  type = string
+}